+++ main.js
@@ -1,315 +1,344 @@
-"use strict"
-const Discord = require("discord.js");
+"use strict";
+
+const Discord = require('discord.js');
 const fs = require('fs');
 const assert = require('assert');
-const _ = require('underscore');
 
-//asrielborg.js modules
+const MAX_WORDLIST_WORD_LENGTH = 20;
+const MIN_WORDLIST_WORD_LENGTH = 1;
+
+/**
+ * Writes a new configuration file from the config object.
+ */
+function writeConfig() {
+    "use strict";
+    fs.writeFile(CONFIG_PATH, JSON.stringify(config, null, ' '), (err) => {
+        if (err) {
+            return log.fatal(`Could not write to ${CONFIG_PATH}. ${err}`);
+        }
+    });
+}
+
+// Asrielborg's modules
 const log = require('./log');
 
-//you can change the paths here, for whatever reason
-const config_path = "./config.json";
-const lines_path = "./lines.txt";
+// You can change the paths here, for whatever reason.
+const CONFIG_PATH = './config.json';
+const LINES_PATH = './lines.txt';
 
-var known_lines = [];
-var known_words = [];
+var linesDictionary = [];
+var wordsDictionary = [];
 var bot;
 
-log.notice("AsrielBorg Version 1.0.1 is now loading ... This might take a while if your lines file is too big.");
+log.notice("Check for updates! They can be found at https://git.io/via60");
+log.notice("AsrielBorg Version 2.0.0 DEV is now loading... This might take a while if your lines file is too big.");
 
 var config = {
-    //The port the panel will run in. TODO
-    //port: 10991,
-    //Discord API Token
-    token: "YOUR TOKEN HERE",
-    //The chance that the bot will reply to any message, in percent.
-    replyrate: 1.0,
-    //The chance that the bot will reply when its nick is mentioned, in percent.
-    replynick: 100.0,
-    //The chance that the bot will reply when a magic word is said, in percent.
-    replymagic: 10.0,
-    
-    //Whether the bot is speaking or not. (0: not learning, 1: learning)
-    speaking: 1,
-    //Whether the bot is learning or not. (0: not learning, 1: learning)
-    learning: 1,
-    
-    //Auto save period (in seconds)
-    autosaveperiod: 200,
-    //The list of magic words the bot will reply to (separated by spaces)
-    magicwords: "this will trigger me",
-    //The list of words that will make the bot not learn a sentence if it contains one of these words
-    blacklistedwords: "very bad word"
+    port: 10991,                        // The port the panel will run in.
+    token: "YOUR TOKEN HERE",           // Discord API Token
+
+    replyRate: 1.0,                     // The chance that the bot will reply to any message, in percent.
+    replyNick: 100.0,                   // The chance that the bot will reply when its nick is mentioned, in percent.
+    replyMagic: 10.0,                   // The chance that the bot will reply when a magic word is said, in percent.
+
+    speaking: true,                     // Whether the bot is speaking or not.
+    learning: true,                     // Whether the bot is learning or not.
+
+    autoSavePeriod: 200,                // Auto save period (in seconds)
+
+    magicWords: ["a trigger", "another trigger"], // The list of magic words the bot will reply to (separated by spaces)
+    blacklistedWords: ["very bad word", "another bad word"]  // The list of words that will make the bot not learn a sentence if it contains one of these words
 };
 
-//Does a config file already exist? Default is config.json.
+// Does a config file already exist? Attempting to load from CONFIG_PATH.
 try {
-    fs.accessSync(config_path, fs.FS_OK);
-    fs.readFile(config_path, (err, data) => {
-        if (err) throw err;
-        validateconf( JSON.parse(data) );
+    fs.accessSync(CONFIG_PATH, fs.FS_OK);
+    fs.readFile(CONFIG_PATH, (err, data) => {
+        if (err) {
+            throw err;
+        }
+        validateConfig(JSON.parse(data));
     });
 } catch (err) {
-    //Config file doesn't exist or is not accessible. Attempt creating a new one.
-    fs.writeFile(config_path, JSON.stringify(config, null, '\n'), (err) => {
+    // Config file doesn't exist or is not accessible. Attempt creating a new one.
+    writeConfig();
+}
+
+/**
+ * Validates the configuration inside the data object.
+ * @param {Object} data - The configuration dictionary.
+ *
+ * Data contains the following fields:
+ * {Number} port                - The port that the panel will run in.
+ * {String} token               - The Discord API token to be used for the bot.
+ * {Number} replyRate           - The chance that the bot will reply to any message, in percent.
+ * {Number} replyMagic          - The chance that the bot will reply when a magic word is said, in percent.
+ * {Number} speaking            - Whether the bot is speaking or not. (0: not learning, 1: learning)
+ * {Number} learning            - Whether the bot is learning or not. (0: not learning, 1: learning)
+ * {Number} autoSavePeriod      - Auto save period (in seconds)
+ * {String} magicWords          - The list of magic words the bot will reply to (separated by spaces).
+ * {String} blacklistedWords    - (Optional) The list of words that will make the bot not learn a sentence if it contains one of these words
+ */
+function validateConfig(data) {
+    "use strict";
+    assert((typeof data.port === 'number' && data.port >= 1 && data.port <= 65535));
+    assert((typeof data.token === 'string'));
+    //
+    assert((typeof data.replyRate === 'number' && data.replyRate >= 0 && data.replyRate <= 100));
+    assert((typeof data.replyNick === 'number' && data.replyNick >= 0 && data.replyNick <= 100));
+    assert((typeof data.replyMagic === 'number' && data.replyMagic >= 0 && data.replyMagic <= 100));
+    //
+    assert((typeof data.speaking === 'boolean'));
+    //
+    assert((typeof data.learning === 'boolean'));
+    //
+    assert((typeof data.autoSavePeriod === 'number' && data.autoSavePeriod >= 0 ));
+    //
+    assert((data.magicWords instanceof Array));
+    data.magicWords = data.magicWords.map((word) => word.toLowerCase().trim());
+    //
+    assert((data.blacklistedWords instanceof Array));
+    data.blacklistedWords = data.blacklistedWords.map((word) => word.toLowerCase().trim());
+
+    config = data; // Save the configuration into memory.
+    loadLinesFile();
+
+}
+
+/**
+ * Load the file in LINES_PATH.
+ * If the file doesn't exist, it will create a new one and then call connect().
+ * If the file does exit,
+ */
+function loadLinesFile() {
+    "use strict";
+    if (!fs.existsSync(LINES_PATH)) {
+        fs.writeFile(LINES_PATH, "", (err) => {
+            if (!err) {
+                log.notice(`Created new ${LINES_PATH}.`);
+                connect(); // Loading was successful; connect to server.
+            } else {
+                console.log(`Could not write to ${LINES_PATH}. ${err}`);
+            }
+        });
+    }
+    fs.readFile(LINES_PATH, (err, data) => {
         if (err) {
-            return log.fatal(`Could not write to ${config_path}. ${err}`);
+            throw err;
         }
-        log.notice(`Written to configuration to ${config_path}. Please check and modify it.`);
+        // data: the file's contents turned into lowercase, with all the carriage returns removed.
+        data = data.toString()
+            .toLowerCase()
+            .replace(/\r/gm, '');
+
+        linesDictionary = data.split(/(\n|\.\ )/)
+            .filter((el) => {
+                return el !== "\n"; // Remove all the unnecessary line breaks.
+            });
+        wordsDictionary = unique(extractWords(data));
+
+        log.notice(`I know ${linesDictionary.length - 1} lines and ${wordsDictionary.length - 1} unique words.`);
+        connect();
     });
 }
 
-function validateconf(data) {
-    //assert((typeof data.port === "number" //TODO
-    //            && data.port >= 1
-    //            && data.port <= 65565
-    //       ));
-    
-    assert((typeof data.token === "string"));
-    
-    assert((typeof data.replyrate === "number"
-                && data.replyrate >= 0 
-                && data.replyrate <= 100
-           ));
-    
-    assert((typeof data.replynick      === "number"
-                && data.replynick >= 0 
-                && data.replynick <= 100
-           ));
-    
-    assert((typeof data.replymagic === "number" 
-                && data.replymagic >= 0 
-                && data.replymagic <= 100
-           ));
-    
-    assert((typeof data.speaking === "number"
-               && (data.speaking == 0 || data.speaking == 1)
-           ));
-    
-    assert((typeof data.learning === "number"
-               && (data.learning == 0 || data.learning == 1)
-           ));
-    
-    assert((typeof data.autosaveperiod === "number"
-                && data.autosaveperiod >= 0
-           ));
-    
-    assert((typeof data.magicwords === "string"));
-    data.magicwords = data.magicwords
-						  .toLowerCase()
-						  .split(' ');
-    
-    assert((typeof data.blacklistedwords === "string" || data.blacklistedwords === undefined));
-    if (data.blacklistedwords === undefined) {
-        data.blacklistedwords = [];
-    } else {
-        data.blacklistedwords = data.blacklistedwords
-                                    .toLowerCase()
-                                    .split(' ');
-    }
-    
-    config = data;
-    loadlines();
-}
-
-function loadlines() {
-    //Attempt loading the lines file.
-	if (!fs.existsSync(lines_path)) {
-		fs.writeFile(lines_path, "", (err) => {
-            if (err) {
-                return console.log(`Could not write to ${lines_path}. ${err}`);
-            }
-            log.notice(`Created new ${lines_path}.`);
-            connect(); //Loading was successful; connect to server.
-        });
-	}
-	fs.readFile(lines_path, (err, data) => {
-		if (err) throw err;
-		
-		let filtereddata = data.toString()
-							 .toLowerCase()
-							 .replace(/(\r)/gm,"")
-		
-		known_lines = filtereddata.split(/(\n|\. )/)
-								.filter((el) => {
-									return el !== "\n"; //remove all the new lines
-								});
-						  
-		known_words = unique(getWords(filtereddata));
-		
-		log.notice(`I know ${known_lines.length - 1} lines and ${known_words.length - 1} unique words.`);
-		connect();
-	});
-}
-
+/**
+ * Connects to Discord using the discord.js API.
+ */
 function connect() {
+    "use strict";
+
     bot = new Discord.Client();
-    
-	bot.login(config.token);
-    
+    bot.login(config.token);
+
     bot.on("ready", () => {
         log.notice("Successfully connected to Discord.");
     });
-    
-    bot.on("message", (message_t) => {
-        if (message_t.author != bot.user) {
-			log.message(message_t.author, message_t.channel, message_t.content);
-            process_message(message_t);
+
+    bot.on("message", (discordMessage) => {
+        log.message(discordMessage.author, discordMessage.channel, discordMessage.content);
+        if (discordMessage.author !== bot.user) {
+            processMessage(discordMessage);
         }
     });
-    
-    bot.on("messageDelete", (message_t) => {
-        message_t.channel.sendMessage(`${message_t.author}: i know what you did`);
+
+    bot.on("messageDelete", (discordMessage) => {
+        discordMessage.channel.sendMessage(`${discordMessage.author}: i know what you did`);
     });
-    
-    setInterval(save_lines, config.autosaveperiod * 1000);
+
+    setInterval(saveLinesFile, config.autoSavePeriod * 1000);
 }
 
-function process_message(message_t) {
+/**
+ * Handles all learning and speech of the bot.
+ * @param {Object} discordMessage - Discord messageString object to be processed
+ */
+function processMessage(discordMessage) {
+    "use strict";
+    let messageString = discordMessage.content.toLowerCase();  // Stores the contents of the messageString object in a string variable.
+
+    if (config.learning) {
+        learn(messageString);
+    }
+
     if (config.speaking) {
-        let message = message_t.content.toLowerCase();
-		
-        let words = getWords(message);
-        
-        if (config.learning) {
-            learn(message);
-        }
-        
+        let words = extractWords(messageString);
+
         if (config.speaking) {
-            let replyflag = false;
+            /**
+             * containsMagic is an anonymous function that will return true if the messageString
+             * contains any of the words inside config.magicWords. Not case sensitive.
+             */
+            let containsMagic = function () {
+                return config.magicWords.find((entry) => messageString.includes(entry));
+            };
 
-            //check if the bot will reply to a magic word
-            let contains_magic = function () {
-                var matching_words = [];
-                config.magicwords.forEach( (magic_word) => {
-                    if (containsCaseInsensitive(words, magic_word)) {
-                        matching_words.push(magic_word);
-                    }
-                });
-                return matching_words.length > 0;
-            }
-            if (contains_magic() && Math.floor(Math.random() * 100) < config.replymagic ) { 
-                replyflag = true;
-            }
-            //check if the bot will reply to nick
-            else if (containsCaseInsensitive(words, bot.user.username)
-                     && Math.floor(Math.random() * 100) < config.replynick) {  
-						replyflag = true;
-					} 
-            //check if the bot will reply to default chance
-            else if (Math.floor(Math.random() * 100) < config.replyrate) { 
-				replyflag = true;
-			}
-            
-            if (replyflag) {
-                reply_to(message_t);
+            let chance = Math.floor(Math.random() * 100);
+
+            if ((containsMagic() && chance <= config.replyMagic)
+                || (messageString.includes(bot.user.username) && chance <= config.replyNick)
+                || (chance <= config.replyRate)) {
+                replyTo(discordMessage);
             }
         }
     }
 }
 
-function reply_to(message_t) {
-    var message = message_t.content.toLowerCase();
-    var words = getWords(message);
-    
-    //search for words that the bot already knows 
-    var known = [];
-    words.forEach( (word, i, arr) => {
-        if (known_words.indexOf(word) > -1) {
-            known.push(word);
+/**
+ * Replies to the messageString object using the Discord API.
+ * @param {Object} discordMessage - The messageString that it will reply to.
+ */
+function replyTo(discordMessage) {
+    "use strict";
+    let messageString = discordMessage.content.toLowerCase(); // Stores the contents of the messageString object in a string variable.
+    let messageWords = extractWords(messageString);
+
+    let recognizedWords = []; // Array of words that the bot recognizes in the messageString.
+    messageWords.forEach((word) => {
+        if (wordsDictionary.includes(word)) {
+            recognizedWords.push(word);
         }
     });
-    
-    if (known.length > 0) {
-        var picked_word = chooseFrom(known);
-        var form_sentence = [picked_word];
-        
-        var re = new RegExp(`\\b${RegExp.quote(picked_word)}\\b`, "gi");
-        
-        //fill left
-        var randomline = get_random_line_containing(picked_word);
-		
-        form_sentence.unshift(randomline.split(re)
-										.splice(0, 1)
-										.join(re)
-							 );
-        //fill right
-        var randomline = get_random_line_containing(picked_word);
-        form_sentence.push(randomline.split(re)
-									 .splice(-1, 1)
-									 .join(re)
-						  );
-        
-        message_t.channel.sendMessage(form_sentence.join(''));
+
+    if (recognizedWords.length <= 0) {
+    } else {
+        let startingWord = chooseRandomFromArray(recognizedWords); // Pick a random word from the recognizedWords words to start building a sentence from there.
+        let messageToSend = [startingWord];            //
+
+        let re = new RegExp(`\\b${escapeRegex(startingWord)}\\b`, "gi"); // This expression matches the picked word.
+
+        // Start filling the output messageString from the left.
+        let randomLine = retrieveRandomLineContaining(startingWord);
+        messageToSend.unshift(randomLine.split(re)
+            .splice(0, 1)
+            .join(re));
+
+        // Start filling the output messageString from the right.
+        randomLine = retrieveRandomLineContaining(startingWord);
+        messageToSend.push(randomLine.split(re)
+            .splice(-1, 1)
+            .join(re));
+
+        discordMessage.channel.sendMessage(messageToSend.join(''));
     }
 }
 
-function learn(message) {
-    var filterlines = message.toLowerCase()
-                             .split(/\. |\n/);
-	
-    let prohibited = false; //Flag that determines whether the sentence contains a blacklisted word
-    if (config.blacklistedwords.length > 0) {
-        filterlines.forEach( (sentence) => {
-            let words = sentence.split(/[\n\.\!\,\;\:\(\)\ \r]/)
-            words.forEach( (word) => {
-                if (config.blacklistedwords.indexOf(word) != -1) {
-                    log.notice(`Refusing to learn sentence containing blacklisted word "${word}".`);
-                    prohibited = true;
-                }
-            });
+/**
+ * @param {String} messageString - The messageString to be learned.
+ */
+function learn(messageString) {
+    "use strict";
+    let choppedMessage = messageString.toLowerCase()
+        .split(/\.\ |\n/);
+
+    let messageWords = extractWords(messageString.toLowerCase());
+
+    let forbiddenWords = []; // Flag that determines whether the sentence contains a blacklisted word
+    if (config.blacklistedWords.length > 0) {
+        config.blacklistedWords.forEach((el) => {
+            if (messageWords.includes(el)) {
+                // Word in messageString matches a blacklisted word.
+                forbiddenWords.push(el);
+            }
         });
     }
-    if (prohibited) { return false; }
-    
-    filterlines.forEach( (sentence) => {
-        if (known_lines.indexOf(sentence) > -1) {
-            //Sentence already found in lines.
-			return false; 
+    if (forbiddenWords.length > 0) {
+        log.notice(`Refusing to learn message containing blacklisted words: ${forbiddenWords.join(', ')}`);
+    }
+
+    choppedMessage.forEach((sentence) => {
+        if (linesDictionary.includes(sentence)) {
+            // Sentence already found in lines.
+            return false;
         }
-        known_lines.push(sentence);
-        
-        let words = getWords(sentence);
-        words.forEach( (word) => {
-            if (known_words.indexOf(word) == -1) {
-                //If word is not known, add word to vocabulary.
-                known_words.push(word);
+        linesDictionary.push(sentence);
+
+        messageWords.forEach((word) => {
+            if (!wordsDictionary.includes(word)) {
+                // Word is not known; add word to vocabulary.
+                wordsDictionary.push(word);
             }
         });
     });
 }
-  
-function save_lines() {
-    fs.writeFile(lines_path, known_lines.join('\r\n'), (err) => {
+
+/**
+ * Write all the bot's lines to the LINES_PATH file.
+ */
+function saveLinesFile() {
+    "use strict";
+    fs.writeFile(LINES_PATH, linesDictionary.join('\r\n'), (err) => {
         if (err) {
-            return log.error(`Could not write to ${lines_path}. ${err}`);
+            return log.error(`Could not write to ${LINES_PATH}. ${err}`);
         }
-        log.notice(`Saved lines at ${Date()}.`);
+        log.notice(`Saved lines at ${new Date()}.`);
     });
 }
 
-//Assist function
-function get_random_line_containing(word) {
-    var line_list = [];
-    var re = new RegExp(`\\b${RegExp.quote(word)}\\b`, 'gi');
-    
-    known_lines.forEach( (line, i, arr) => {
-        if ( line.match(re) ) {
-            line_list.push(line);
-        }
-    });
-    return line_list.length > 0 ? chooseFrom(line_list) : "";
+/**
+ * Choose a random element from an array.
+ * @param {Array} array
+ * @return {Object} - random object within the array
+ */
+function chooseRandomFromArray(array) {
+    "use strict";
+    if (array.length === 0) {
+        return null;
+    }
+    return array[Math.floor(Math.random() * array.length)];
 }
 
-//Assist function
-//Removes all duplicates from an array
+/**
+ * Returns a random line that contains a word.
+ * @param {String} word - The word to search for.
+ * @return {String} - The random line found in the known lines.
+ */
+function retrieveRandomLineContaining(word) {
+    "use strict";
+    let matchingLines = [];
+    let re = new RegExp(`\\b${escapeRegex(word)}\\b`, 'gi');
+
+    linesDictionary.forEach((line, i, arr) => {
+        if (line.match(re)) {
+            matchingLines.push(line);
+        }
+    });
+    return matchingLines.length > 0 ? chooseRandomFromArray(matchingLines) : "";
+}
+
+/**
+ * Removes all duplicates from a provided array.
+ * @param {Array} array
+ * @return {Array} - The processed array.
+ */
 function unique(array) {
-    var dict = {};
-    var i;
-    var l = array.length;
-    var r = [];
+    "use strict";
+    let dict = {};
+    let i;
+    let l = array.length;
+    let r = [];
     for (i = 0; i < l; i += 1) {
-        dict[array[i]] = array[i]; 
+        dict[array[i]] = array[i];
     }
     for (i in dict) {
         r.push(dict[i]);
@@ -317,38 +346,22 @@ function unique(array) {
     return r;
 }
 
-//Assist function
-function isChannel(name) {
-    if ( /^[#&]\w+$/.test(name) ) {
-        return true;
-    return false;
-    }
+/**
+ * Allows use of variable strings in regular expressions.
+ * @param {String} str
+ * @return {String}
+ */
+function escapeRegex(str) {
+    "use strict";
+    return str.replace(/([.?*+\^$\[\]\\(){}\|\-])/g, "\\$1");
 }
 
-
-//Assist function
-RegExp.quote = function(str) {
-     return str.replace(/([.?*+^$[\]\\(){}|-])/g, "\\$1");
-};
-
-//Assist function
-function containsCaseInsensitive(array1, element_) {
-    return array1.some( function (el, i) { 
-        if (el.toLowerCase() === element_.toLowerCase()) {
-            return true;
-        }
-    });
-}
-    
-//Assist function
-function chooseFrom(array1) {
-    if (array1.length == 0) {
-        return [];
-    } 
-    return array1[Math.floor(Math.random() * array1.length)];
-}
-
-//Assist function
-function getWords(line) {
-    return line.split(/[\.\!\,\;\:\(\)\ \?]/);
+/**
+ * Filter the words from a line string.
+ * @param {String} line
+ * @return {Array} - The array containing the words of the string.
+ */
+function extractWords(line) {
+    "use strict";
+    return line.split(/[.!,;:()?\ ]/);
 }