+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+[1.0.2] - 2016-09-18
+   - Magic/Blacklisted words list in config.json are now an array of strings.
+   - camelCased everything
+   - Lots of cleanup
+
 [1.0.1] - 2016-09-10
    - Added a blacklist field in ```config.json```. ```"blacklistedwords": "example words go like this"```
    - Some changes to strict usage of variables